Modified stoch.cpp to eliminate "scrambling" matrix elements. Added some messages to ScnModel.readSmps

+++ Smi/src/SmiScenarioTree.hpp
@@ -224,7 +224,7 @@ public:
 	/** get number of scenarios */
 	int getNumScenarios()
 	{
-		return leaf_.size();
+		return (int) leaf_.size();
 	}
